Move linting to linters folder for testing
